Crate jadis

Introduction

The Jadis crate is a Java Virtual Machine language disassembler written in Rust. This project has been written for educational purposes, however, Jadis can theoretically be used as a drop-in replacement of javap.

Using Jadis

It is fairly straightforward to use Jadis. Simply run the project and use the same command-line arguments you would use with javap. Please do note that the output format could be slightly different than the output you are used to. Having said that, though, the data in the output should be identical to the output of the disassembler that came with your JDK.

option	description

–bootclasspath	Override location of bootstrap class files

–class-path	Specify where to find user class files

–classpath	Specify where to find user class files

-c	Disassemble the code

–constants	Show final constants

–cp	Specify where to find user class files

-h, –help	Print this help message

-J	Specify a VM option

-l	Print line number and local variable tables

-m, –module	Specify module containing classes to be disassembled

–module-path	Specify where to find application modules

–multi-release	Specify the version to use in multi-release JAR files

–package	Show package/protected/public classes and members (default)

-p, –private	Show all classes and members

–protected	Show protected/public classes and members

–public	Show only public classes and members

-s	Print internal type signatures

–sysinfo	Show system info (path, size, date, SHA-256 hash) of class being processed

–system	Specify where to find system modules

-V, –version	Version information

-v, –verbose	Print additional information

Modules

access_flags
Contains class access and property modifiers

byte_reader
Simplifies reading bytes from binary files

class_file
Represents the JVM .class format

constant_pool
Represents an entity in the constant pool section of a class file

disassembler
Java Virtual Machine language disassembler

utils
Contains useful common functionality and utilities

Functions

main
Application entry point