Module jadis::attribute

Provides functionality to simplify working with attribute structures

Reference: https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-4.html#jvms-4.7

Structs

AttributeAnnotationDefault

AttributeBootstrapMethods
Records bootstrap methods used to produce dynamically-computed constants and dynamically-computed call sites https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-4.html#jvms-4.7.23

AttributeCode
A code attribute contains the Java Virtual Machine instructions and auxilary information for a method, including an instance initialization method and a class or interface initialization method

AttributeConstantValue
Represents the value of a constant expression

AttributeDeprecated
The deprecated attribute is used to indicate that the class, interface, method, or field has been superseded

AttributeEnclosingMethod
A class must have an enclosing method attribute if and only if it represents a local class or an anonymous class

AttributeExceptions
Exceptions attributes indicate which checked exceptions a method may throw

AttributeInfo
Represents an attribute

AttributeInnerClasses
Used inside a class file structure to provide information about the class or interface

AttributeLineNumberTable
A line number table attribute may be used by debuggers to determine which part of the code array corresponds to a given line number in the original source file

AttributeLocalVariableTable
May be used by debuggers to determine the value of a given local variable during the execution of a method

AttributeLocalVariableTypeTable
May be used by debuggers to determine the value of a given local variable during the execution of a method

AttributeMethodParameters
Records information about the formal parameters of a method, such as their names

AttributeModule
The Module attribute indicates the modules required by a module; the packages exported and opened by a module; and the services used and provided by a module

AttributeModuleMainClass
The ModuleMainClass attribute indicates the main class of a module

AttributeModulePackages
The ModulePackages attribute indicates all the packages of a module that are exported or opened by the Module attribute, as well as all the packages of the service implementations recorded in the Module attribute

AttributeNestHost
The NestHost attribute records the nest host of the nest to which the current class or interface claims to belong

AttributeNestMembers
The NestMembers attribute records the classes and interfaces that are authorized to claim membership in the nest hosted by the current class or interface

AttributePermittedSubclasses
The PermittedSubclasses attribute records the classes and interfaces that are authorized to directly extend or implement the current class or interface

AttributeRecord
The Record attribute indicates that the current class is a record class, and stores information about the record components of the record class

AttributeRuntimeInvisibleAnnotations

AttributeRuntimeInvisibleParameterAnnotations

AttributeRuntimeInvisibleTypeAnnotations

AttributeRuntimeVisibleAnnotations

AttributeRuntimeVisibleParameterAnnotations

AttributeRuntimeVisibleTypeAnnotations

AttributeSignature
A Signature attribute stores a signature for a class, interface, constructor, method, field, or record component whose declaration in the Java programming language uses type variables or parameterized types

AttributeSourceDebugExtension
Holds extended debugging information which has no semantic effect on the Java Virtual Machine

AttributeSourceFile
Source file attributes represent the name of the source file from which this class file was compiled

AttributeStackMapTable

AttributeSynthetic
Synthetic attributes represent class members that do not appear in the source code

BootstrapMethodEntry
Represents a bootstrap method information entry

ExceptionTableEntry
Describes an exception handler in the code array

InnerClassEntry
Represents a class entry in the inner classes attribute

LineNumberTableEntry
Represents an entry in the line number table in a line number table attribute

LocalVariableTableEntry
Indicates a range of code array offsets within which a local variable has a value, and indicates the index into the local variable array of the current frame at which that local variable can be found

LocalVariableTypeTableEntry
Indicates a range of code array offsets within which a local variable has a value, and indicates the index into the local variable array of the current frame at which that local variable can be found

MethodParameterEntry
Represents information about a method parameter

ModuleExportsEntry
Indicates the number of entries in the exports table

ModuleOpensEntry
Specifies a package opened by the current module, such that all types in the package, and all their members, may be accessed from outside the current module via the reflection libraries of the Java SE Platform, possibly from a limited set of “friend” modules.

ModuleProvidesEntry
Represents a service implementation for a given service interface

ModuleRequiresEntry
Specifies a dependence of the current module

RecordComponentInfo
Specifies a record component of the current class

Enums

AttributeType
Attribute types

Traits

Attribute
Base trait to store specialised attributes