Struct jadis::classfile::attribute::AttributeInfo

pub struct AttributeInfo {
    pub attribute_type: AttributeType,
    data: Box<dyn Attribute>,
}

Represents an attribute

Fields
attribute_type: AttributeType

Identifies the type of attribute this structure represents

data: Box<dyn Attribute>

Data associated with this attribute

Implementations

impl AttributeInfo

pub fn new(
reader: &mut ByteReader,
constant_pool: &BTreeMap<u16, ConstantPoolInfo>
) -> Self

Create a new attribute from a class file binary blob

fn read_data_as_constant_value(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeConstantValue

Read the data blob as a constant value attribute

fn read_data_as_code(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32,
constant_pool: &BTreeMap<u16, ConstantPoolInfo>
) -> AttributeCode

Read the data blob as a code attribute

fn read_data_as_stack_map_table(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeStackMapTable

Read the data blob as a stack map table attribute

fn read_data_as_exceptions(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeExceptions

Read the data blob as an exceptions attribute

fn read_data_as_inner_classes(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeInnerClasses

Read the data blob as an inner classes attribute

fn read_data_as_enclosing_method(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeEnclosingMethod

Read the data blob as an enclosing method attribute

fn read_data_as_synthetic(
attribute_name_index: u16,
attribute_length: u32
) -> AttributeSynthetic

Read the data blob as a synthetic attribute

fn read_data_as_signature(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeSignature

Read the data blob as a signature attribute

fn read_data_as_source_file(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeSourceFile

Read the data blob as a source file attribute

fn read_data_as_source_debug_extension(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeSourceDebugExtension

Read the data blob as a source debug extension attribute

fn read_data_as_line_number_table(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeLineNumberTable

Read the data blob as a line number table attribute

fn read_data_as_local_variable_table(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeLocalVariableTable

Read the data blob as a local variable table attribute

fn read_data_as_local_variable_type_table(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeLocalVariableTypeTable

Read the data blob as a local variable type table attribute

fn read_data_as_deprecated(
attribute_name_index: u16,
attribute_length: u32
) -> AttributeDeprecated

Read the data blob as a deprecated attribute

fn read_data_as_runtime_visible_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeVisibleAnnotations

Read the data blob as a runtime visible annotations attribute

fn read_data_as_runtime_invisible_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeInvisibleAnnotations

Read the data blob as a runtime invisible annotations attribute

fn read_data_as_runtime_visible_parameter_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeVisibleParameterAnnotations

Read the data blob as a runtime visible parameter annotations attribute

fn read_data_as_runtime_invisible_parameter_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeInvisibleParameterAnnotations

Read the data blob as a runtime invisible parameter annotations attribute

fn read_data_as_runtime_visible_type_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeVisibleTypeAnnotations

Read the data blob as a runtime visible type annotations attribute

fn read_data_as_runtime_invisible_type_annotations(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeRuntimeInvisibleTypeAnnotations

Read the data blob as a runtime invisible type annotations attribute

fn read_data_as_annotation_default(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeAnnotationDefault

Read the data blob as an annotation default attribute

fn read_data_as_bootstrap_methods(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeBootstrapMethods

Read the data blob as a bootstrap methods attribute

fn read_data_as_method_parameters(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeMethodParameters

Read the data blob as a method parameters attribute

fn read_data_as_module(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeModule

Read the data blob as a module attribute

fn read_data_as_module_packages(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeModulePackages

Read the data blob as a module packages attribute

fn read_data_as_module_main_class(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeModuleMainClass

Read the data blob as a module main class attribute

fn read_data_as_nest_host(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeNestHost

Read the data blob as a nest host attribute

fn read_data_as_nest_members(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributeNestMembers

Read the data blob as a nest members attribute

fn read_data_as_record(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32,
constant_pool: &BTreeMap<u16, ConstantPoolInfo>
) -> AttributeRecord

Read the data blob as a record attribute

fn read_data_as_permitted_subclasses(
reader: &mut ByteReader,
attribute_name_index: u16,
attribute_length: u32
) -> AttributePermittedSubclasses

Read the data blob as a permitted subclasses attribute