Module jadis::constant_pool

Represents an entity in the constant pool section of a class file

This module contains all information necessary to parse constant pool entities from class files

Structs

ConstantClassInfo
Constant pool class

ConstantDoubleInfo
Constant pool double

ConstantDynamicInfo
Constant pool dynamic

ConstantFieldRefInfo
Constant pool field reference

ConstantFloatInfo
Constant pool float

ConstantIntegerInfo
Constant pool integer

ConstantInterfaceMethodRefInfo
Constant pool interface method reference

ConstantInvokeDynamicInfo
Constant pool invoke dynamic

ConstantLongInfo
Constant pool long

ConstantMethodHandleInfo
Constant pool method handle

ConstantMethodRefInfo
Constant pool method reference

ConstantMethodTypeInfo
Constant pool method type

ConstantModuleInfo
Constant pool module

ConstantNameAndTypeInfo
Constant pool name and type

ConstantPackageInfo
Constant pool package

ConstantPoolInfo
Represents an entity in the constant pool

ConstantStringInfo
Constant pool string

ConstantUtf8Info
Constant pool UTF-8 string

Enums

MethodHandleType
Bytecode behaviours for method handles

Tag
Constant pool tags

Traits

ConstantPoolInfoData
Base trait to store specialised constant pool data entries