Struct jadis::constant_pool::ConstantPoolInfo

pub struct ConstantPoolInfo {
    pub tag: Tag,
    data: Box<dyn ConstantPoolInfoData>,
}

Represents an entity in the constant pool

Fields
tag: Tag

Identifies the type of data this entity represents

data: Box<dyn ConstantPoolInfoData>

Data associated with this entity

Implementations

impl ConstantPoolInfo

pub fn new(reader: &mut ByteReader, index: u16) -> Self

Create a new constant pool entity from a class file binary blob

fn read_data_as_utf8(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantUtf8Info

Read the data blob as an UTF-8 constant pool entry

fn read_data_as_integer(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantIntegerInfo

Read the data blob as an integer constant pool entry

fn read_data_as_float(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantFloatInfo

Read the data blob as a float constant pool entry

fn read_data_as_long(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantLongInfo

Read the data blob as a long constant pool entry

fn read_data_as_double(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantDoubleInfo

Read the data blob as a double constant pool entry

fn read_data_as_class(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantClassInfo

Read the data blob as a class constant pool entry

fn read_data_as_string(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantStringInfo

Read the data blob as a string constant pool entry

fn read_data_as_field_ref(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantFieldRefInfo

Read the data blob as a field reference constant pool entry

fn read_data_as_method_ref(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantMethodRefInfo

Read the data blob as a method reference constant pool entry

fn read_data_as_interface_method_ref(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantInterfaceMethodRefInfo

Read the data blob as an interface method reference constant pool entry

fn read_data_as_name_and_type(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantNameAndTypeInfo

Read the data blob as a name and type constant pool entry

fn read_data_as_method_handle(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantMethodHandleInfo

Read the data blob as a method handle constant pool entry

fn read_data_as_method_type(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantMethodTypeInfo

Read the data blob as a method type constant pool entry

fn read_data_as_dynamic(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantDynamicInfo

Read the data blob as a dynamic constant pool entry

fn read_data_as_invoke_dynamic(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantInvokeDynamicInfo

Read the data blob as an invoke dynamic constant pool entry

fn read_data_as_module(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantModuleInfo

Read the data blob as a module constant pool entry

fn read_data_as_package(
reader: &mut ByteReader,
constant_pool_index: u16
) -> ConstantPackageInfo

Read the data blob as a package constant pool entry

pub fn try_cast_into_utf8(&self) -> Option<&ConstantUtf8Info>

Cast to as an UTF-8 constant pool entry

pub fn try_cast_into_integer(&self) -> Option<&ConstantIntegerInfo>

Cast to an integer constant pool entry

pub fn try_cast_into_float(&self) -> Option<&ConstantFloatInfo>

Cast to a float constant pool entry

pub fn try_cast_into_long(&self) -> Option<&ConstantLongInfo>

Cast to a long constant pool entry

pub fn try_cast_into_double(&self) -> Option<&ConstantDoubleInfo>

Cast to a double constant pool entry

pub fn try_cast_into_class(&self) -> Option<&ConstantClassInfo>

Cast to a class constant pool entry

pub fn try_cast_into_string(&self) -> Option<&ConstantStringInfo>

Cast to a string constant pool entry

pub fn try_cast_into_field_ref(&self) -> Option<&ConstantFieldRefInfo>

Cast to a field reference constant pool entry

pub fn try_cast_into_method_ref(&self) -> Option<&ConstantMethodRefInfo>

Cast to a method reference constant pool entry

pub fn try_cast_into_interface_method_ref(
&self
) -> Option<&ConstantInterfaceMethodRefInfo>

Cast to an interface method reference constant pool entry

pub fn try_cast_into_name_and_type(&self) -> Option<&ConstantNameAndTypeInfo>

Cast to a name and type constant pool entry

pub fn try_cast_into_method_handle(&self) -> Option<&ConstantMethodHandleInfo>

Cast to a method handle constant pool entry

pub fn try_cast_into_method_type(&self) -> Option<&ConstantMethodTypeInfo>

Cast to a method type constant pool entry

pub fn try_cast_into_dynamic(&self) -> Option<&ConstantDynamicInfo>

Cast to a dynamic constant pool entry

pub fn try_cast_into_invoke_dynamic(&self) -> Option<&ConstantInvokeDynamicInfo>

Cast to an invoke dynamic constant pool entry

pub fn try_cast_into_module(&self) -> Option<&ConstantModuleInfo>

Cast to a module constant pool entry

pub fn try_cast_into_package(&self) -> Option<&ConstantPackageInfo>

Cast to a package constant pool entry