Trait jadis::constant_pool::ConstantPoolInfoData

pub trait ConstantPoolInfoData {
    fn as_concrete_type(&self) -> &dyn Any;
}

Base trait to store specialised constant pool data entries

Required methods

fn as_concrete_type(&self) -> &dyn Any

Cast to the concreate type that implements this trait